Moved exisiting tests to consensus_engine
