+++ test_integration/consensus_engine/01_deployment/01_deploy.ts
@@ -14,8 +14,8 @@
 
 'use strict';
 
-import shared from '../shared';
-import Interacts from "../../interacts/Interacts";
+import shared from '../../shared';
+import Interacts from "../../../interacts/Interacts";
 
 const FUNDING_AMOUNT_IN_ETHER = '2';
 describe('Deployment', async () => {
